// packages/plugin-calendar-ui/src/settings/types.ts
export interface ICalendar {
  _id: string;
  name: string;
  color: string;
  groupId: string;
  accountId: string;
  userId: string;
}

export interface IGroup {
  _id: string;
  name: string;
  isPrivate: boolean;
  boardId: string;
  memberIds: string[];
  calendars: ICalendar[];
}